+++ CHANGELOG.md
@@ -7,8 +7,10 @@ GAMSPy 1.0.4
   - Do not create a GDX file when it's not necessary. 
   - Do not carry solver options from the previous solve to the new solve.
   - Fix toGams bug of MathOp symbols.
+  - Add "same" and "valid" options for Conv2d padding.
 - Testing
   - Add one to one comparison tests with reference files in toGams tests.
+  - Add tests for "same" and "valid" padding options of Conv2d.
 
 GAMSPy 1.0.3
 ------------