+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 GAMSPy CHANGELOG
 ================
 
+GAMSPy 1.0.5
+------------
+- General
+  - Allow printing the records of variable/equation attributes with a print(variable.attribute[domain].records) syntax.
+  - Allow printing the records of a subset of a parameter with print(parameter['literal'].records) syntax.
+
 GAMSPy 1.0.4
 ------------
 - General